Updated Helpers (ushort->uint version)
Forgot to include this change on the previous commit, whoops!

+++ Tools/Helpers.cs
@@ -9,7 +9,7 @@ namespace MapleShark
 {
     class Helpers
     {
-        public static string GetScriptFolder(byte locale, ushort build)
+        public static string GetScriptFolder(byte locale, uint build)
         {
             return string.Format(
                 "{1}{0}Scripts{0}{2}{0}{3}{0}", 
@@ -20,7 +20,7 @@ namespace MapleShark
             );
         }
 
-        public static string GetScriptPath(byte locale, ushort build, bool outbound, ushort opcode)
+        public static string GetScriptPath(byte locale, uint build, bool outbound, ushort opcode)
         {
             return string.Format(
                 "{1}{2}{0}0x{3:X4}.txt",
@@ -31,7 +31,7 @@ namespace MapleShark
             );
         }
 
-        public static string GetCommonScriptPath(byte locale, ushort build)
+        public static string GetCommonScriptPath(byte locale, uint build)
         {
             return Path.Combine(GetScriptFolder(locale, build), "Common.txt");
         }
